test(aspect): Ensure we are testing with the desired Bazel version
Since, we discovered the Windows CI tests for the examples are not behaving
as desired, we should also harden our aspect test setups.
