+++ mypy.ini
@@ -12,4 +12,4 @@ warn_unused_ignores = true
 warn_no_return = true
 warn_unreachable = true
 # Enable relative imports in the test orchestration scripts
-mypy_path = test/aspect/:test/apply_fixes/
+mypy_path = test/:test/aspect/:test/apply_fixes/
